refactor(bottomtab): moved code to separate file + updated references
